+++ src/data_analysis_gui/widgets/__init__.py
@@ -7,7 +7,6 @@ License: MIT (see LICENSE file for details)
 from .custom_inputs import (
     SelectAllLineEdit,
     SelectAllSpinBox,
-    SelectAllIntSpinBox,
     NoScrollComboBox,
 )
 from data_analysis_gui.widgets.concentration_range_table import ConcentrationRangeTable
@@ -15,7 +14,6 @@ from data_analysis_gui.widgets.concentration_range_table import ConcentrationRan
 __all__ = [
     "SelectAllLineEdit",
     "SelectAllSpinBox",
-    "SelectAllIntSpinBox",
     "NoScrollComboBox",
     "ConcentrationRangeTable",
 ]